Remove old ast classes
